updated min fare test

+++ src/main/java/io/ljn/jp/test/runner/journey/FarePrice.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class FarePrice {
-    public final double price;
+    public final float price;
 }
